Changed specification for SSH key structure
Previously, the keydir contained keys of the following format:

    `owner/owner.pub`
or

    `owner/location/owner.pub`

Now, we assert a default location for keys (defined in the settings)
for all keys below the root folder.
Additionally, we added a `subfolders` attribute to define
paths below the keydir for better organization.

This allows the following key formats (relative to <key_dir>)

    1. `owner.pub` (where domain is a name or email)
    2. `location/owner.pub`
    3. `some/subfolder/location/owner.pub`
    4. `owner/location/owner.pub` (With owner as subfolders)

The previously possible, legacy format:

    `owner/owner.pub`

will be read, but the location set to the default
(c.f., defaults `location`) and written to:

    `owner/<default_location>/owner.pub`